include/bits.h: Change getleu64() and getles64() to C99 units.
Too many issues with uint64_t and int64_t.
Also remove to unused #defines from navcom.c:
getles64_be() and getleu64_be

+++ include/bits.h
@@ -36,8 +36,8 @@
 #define getleu16(buf, off)      ((uint16_t)(((uint16_t)getub((buf), (off)+1) << 8) | (uint16_t)getub((buf), (off))))
 #define getles32(buf, off)      ((int32_t)(((uint16_t)getleu16((buf),  (off)+2) << 16) | (uint16_t)getleu16((buf), (off))))
 #define getleu32(buf, off)      ((uint32_t)(((uint16_t)getleu16((buf),(off)+2) << 16) | (uint16_t)getleu16((buf), (off))))
-#define getles64(buf, off)      ((int64_t)(((uint64_t)getleu32(buf, (off)+4) << 32) | getleu32(buf, (off))))
-#define getleu64(buf, off)      ((uint64_t)(((uint64_t)getleu32(buf, (off)+4) << 32) | getleu32(buf, (off))))
+#define getles64(buf, off)      ((long long)(((unsigned long long)getleu32(buf, (off)+4) << 32) | getleu32(buf, (off))))
+#define getleu64(buf, off)      ((unsigned long long)(((unsigned long long)getleu32(buf, (off)+4) << 32) | getleu32(buf, (off))))
 extern float getlef32(const char *, int);
 extern double getled64(const char *, int);
